feat(project): 添加项目动态记录功能
- 将ProjectRecord实体中的targetId字段从Integer改为Long类型
- 在ProjectRecordEvent构造函数中初始化targetId为null
- 新增ProjectRecordListener事件监听器处理项目动态记录事件
- 在ProjectRecordServiceImpl中添加addProjectRecord方法
- 扩展ProjectRecordType枚

+++ junoyi-module/junoyi-module-project/src/main/java/com/junoyi/project/domain/po/ProjectRecord.java
@@ -44,7 +44,7 @@ public class ProjectRecord {
     /**
      * 操作目标ID
      */
-    private Integer targetId;
+    private Long targetId;
 
     /**
      * 操作内容
